+++ client/rest/src/main/java/org/elasticsearch/client/RestClient.java
@@ -27,7 +27,6 @@ import org.apache.http.HttpHost;
 import org.apache.http.HttpRequest;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.ClientProtocolException;
-import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.HttpEntityEnclosingRequestBase;
 import org.apache.http.client.methods.HttpHead;
 import org.apache.http.client.methods.HttpOptions;
@@ -39,12 +38,8 @@ import org.apache.http.client.methods.HttpTrace;
 import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.concurrent.FutureCallback;
 import org.apache.http.entity.ContentType;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.impl.nio.client.CloseableHttpAsyncClient;
-import org.apache.http.impl.nio.client.HttpAsyncClientBuilder;
 import org.apache.http.nio.client.methods.HttpAsyncMethods;
-import org.apache.http.nio.conn.SchemeIOSessionStrategy;
 import org.apache.http.nio.protocol.HttpAsyncRequestProducer;
 import org.apache.http.nio.protocol.HttpAsyncResponseConsumer;
 
@@ -70,7 +65,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Client that connects to an elasticsearch cluster through http.
- * Must be created using {@link Builder}, which allows to set all the different options or just rely on defaults.
+ * Must be created using {@link RestClientBuilder}, which allows to set all the different options or just rely on defaults.
  * The hosts that are part of the cluster need to be provided at creation time, but can also be replaced later
  * by calling {@link #setHosts(HttpHost...)}.
  * The method {@link #performRequest(String, String, Map, HttpEntity, Header...)} allows to send a request to the cluster. When
@@ -509,7 +504,6 @@ public final class RestClient implements Closeable {
         public void onSuccess(Response response) {
             this.response = response;
             latch.countDown();
-
         }
 
         @Override
@@ -530,154 +524,10 @@ public final class RestClient implements Closeable {
     }
 
     /**
-     * Returns a new {@link Builder} to help with {@link RestClient} creation.
+     * Returns a new {@link RestClientBuilder} to help with {@link RestClient} creation.
      */
-    public static Builder builder(HttpHost... hosts) {
-        return new Builder(hosts);
-    }
-
-    /**
-     * Helps creating a new {@link RestClient}. Allows to set the most common http client configuration options when internally
-     * creating the underlying {@link org.apache.http.nio.client.HttpAsyncClient}. Also allows to provide an externally created
-     * {@link org.apache.http.nio.client.HttpAsyncClient} in case additional customization is needed.
-     */
-    public static final class Builder {
-        public static final int DEFAULT_CONNECT_TIMEOUT_MILLIS = 1000;
-        public static final int DEFAULT_SOCKET_TIMEOUT_MILLIS = 10000;
-        public static final int DEFAULT_MAX_RETRY_TIMEOUT_MILLIS = DEFAULT_SOCKET_TIMEOUT_MILLIS;
-        public static final int DEFAULT_CONNECTION_REQUEST_TIMEOUT_MILLIS = 500;
-        public static final int DEFAULT_MAX_CONN_PER_ROUTE = 10;
-        public static final int DEFAULT_MAX_CONN_TOTAL = 30;
-
-        private static final Header[] EMPTY_HEADERS = new Header[0];
-
-        private final HttpHost[] hosts;
-        private int maxRetryTimeout = DEFAULT_MAX_RETRY_TIMEOUT_MILLIS;
-        private Header[] defaultHeaders = EMPTY_HEADERS;
-        private FailureListener failureListener;
-        private HttpClientConfigCallback httpClientConfigCallback;
-        private RequestConfigCallback requestConfigCallback;
-
-        /**
-         * Creates a new builder instance and sets the hosts that the client will send requests to.
-         */
-        private Builder(HttpHost... hosts) {
-            if (hosts == null || hosts.length == 0) {
-                throw new IllegalArgumentException("no hosts provided");
-            }
-            this.hosts = hosts;
-        }
-
-        /**
-         * Sets the default request headers, which will be sent along with each request
-         */
-        public Builder setDefaultHeaders(Header[] defaultHeaders) {
-            Objects.requireNonNull(defaultHeaders, "defaultHeaders must not be null");
-            for (Header defaultHeader : defaultHeaders) {
-                Objects.requireNonNull(defaultHeader, "default header must not be null");
-            }
-            this.defaultHeaders = defaultHeaders;
-            return this;
-        }
-
-        /**
-         * Sets the {@link FailureListener} to be notified for each request failure
-         */
-        public Builder setFailureListener(FailureListener failureListener) {
-            Objects.requireNonNull(failureListener, "failureListener must not be null");
-            this.failureListener = failureListener;
-            return this;
-        }
-
-        /**
-         * Sets the maximum timeout (in milliseconds) to honour in case of multiple retries of the same request.
-         * {@link #DEFAULT_MAX_RETRY_TIMEOUT_MILLIS} if not specified.
-         *
-         * @throws IllegalArgumentException if maxRetryTimeoutMillis is not greater than 0
-         */
-        public Builder setMaxRetryTimeoutMillis(int maxRetryTimeoutMillis) {
-            if (maxRetryTimeoutMillis <= 0) {
-                throw new IllegalArgumentException("maxRetryTimeoutMillis must be greater than 0");
-            }
-            this.maxRetryTimeout = maxRetryTimeoutMillis;
-            return this;
-        }
-
-        /**
-         * Sets the {@link HttpClientConfigCallback} to be used to customize http client configuration
-         */
-        public Builder setHttpClientConfigCallback(HttpClientConfigCallback httpClientConfigCallback) {
-            Objects.requireNonNull(httpClientConfigCallback, "httpClientConfigCallback must not be null");
-            this.httpClientConfigCallback = httpClientConfigCallback;
-            return this;
-        }
-
-        /**
-         * Sets the {@link RequestConfigCallback} to be used to customize http client configuration
-         */
-        public Builder setRequestConfigCallback(RequestConfigCallback requestConfigCallback) {
-            Objects.requireNonNull(requestConfigCallback, "requestConfigCallback must not be null");
-            this.requestConfigCallback = requestConfigCallback;
-            return this;
-        }
-
-        /**
-         * Creates a new {@link RestClient} based on the provided configuration.
-         */
-        public RestClient build() {
-            if (failureListener == null) {
-                failureListener = new FailureListener();
-            }
-            CloseableHttpAsyncClient httpClient = createHttpClient();
-            httpClient.start();
-            return new RestClient(httpClient, maxRetryTimeout, defaultHeaders, hosts, failureListener);
-        }
-
-        private CloseableHttpAsyncClient createHttpClient() {
-            //default timeouts are all infinite
-            RequestConfig.Builder requestConfigBuilder = RequestConfig.custom().setConnectTimeout(DEFAULT_CONNECT_TIMEOUT_MILLIS)
-                    .setSocketTimeout(DEFAULT_SOCKET_TIMEOUT_MILLIS)
-                    .setConnectionRequestTimeout(DEFAULT_CONNECTION_REQUEST_TIMEOUT_MILLIS);
-            if (requestConfigCallback != null) {
-                requestConfigCallback.customizeRequestConfig(requestConfigBuilder);
-            }
-
-            HttpAsyncClientBuilder httpClientBuilder = HttpAsyncClientBuilder.create().setDefaultRequestConfig(requestConfigBuilder.build())
-                    //default settings for connection pooling may be too constraining
-                    .setMaxConnPerRoute(DEFAULT_MAX_CONN_PER_ROUTE).setMaxConnTotal(DEFAULT_MAX_CONN_TOTAL);
-            if (httpClientConfigCallback != null) {
-                httpClientConfigCallback.customizeHttpClient(httpClientBuilder);
-            }
-            return httpClientBuilder.build();
-        }
-    }
-
-    /**
-     * Callback used the default {@link RequestConfig} being set to the {@link CloseableHttpClient}
-     * @see HttpClientBuilder#setDefaultRequestConfig
-     */
-    public interface RequestConfigCallback {
-        /**
-         * Allows to customize the {@link RequestConfig} that will be used with each request.
-         * It is common to customize the different timeout values through this method without losing any other useful default
-         * value that the {@link RestClient.Builder} internally sets.
-         */
-        void customizeRequestConfig(RequestConfig.Builder requestConfigBuilder);
-    }
-
-    /**
-     * Callback used to customize the {@link CloseableHttpClient} instance used by a {@link RestClient} instance.
-     * Allows to customize default {@link RequestConfig} being set to the client and any parameter that
-     * can be set through {@link HttpClientBuilder}
-     */
-    public interface HttpClientConfigCallback {
-        /**
-         * Allows to customize the {@link CloseableHttpAsyncClient} being created and used by the {@link RestClient}.
-         * Commonly used to customize the default {@link org.apache.http.client.CredentialsProvider} for authentication
-         * or the {@link SchemeIOSessionStrategy} for communication through ssl without losing any other useful default
-         * value that the {@link RestClient.Builder} internally sets, like connection pooling.
-         */
-        void customizeHttpClient(HttpAsyncClientBuilder httpClientBuilder);
+    public static RestClientBuilder builder(HttpHost... hosts) {
+        return new RestClientBuilder(hosts);
     }
 
     /**
